Rename portable barrel to universal
