+++ .changeset/universal-chart-barrels.md
@@ -2,5 +2,5 @@
 '@tanstack/charts': minor
 ---
 
-Add environment-safe `/portable` and `/types` entry points while preserving
+Add environment-safe `/universal` and `/types` entry points while preserving
 the existing browser-oriented root exports.
