+++ packages/charts-core/package.json
@@ -54,7 +54,7 @@
     "./link": "./src/link.ts",
     "./mark/scale-values": "./src/mark-with-scale-values.ts",
     "./polar": "./src/polar.ts",
-    "./portable": "./src/portable.ts",
+    "./universal": "./src/universal.ts",
     "./reconcile": "./src/reconcile.ts",
     "./rect": "./src/rect.ts",
     "./renderer": "./src/renderer.ts",
@@ -68,7 +68,7 @@
     "./tick": "./src/tick.ts",
     "./tooltip": "./src/tooltip.ts",
     "./tooltip/portal": "./src/tooltip-portal.ts",
-    "./types": "./src/portable-types.ts",
+    "./types": "./src/universal-types.ts",
     "./vector": "./src/vector.ts"
   },
   "publishConfig": {
@@ -171,9 +171,9 @@
         "types": "./dist/polar.d.ts",
         "import": "./dist/polar.js"
       },
-      "./portable": {
-        "types": "./dist/portable.d.ts",
-        "import": "./dist/portable.js"
+      "./universal": {
+        "types": "./dist/universal.d.ts",
+        "import": "./dist/universal.js"
       },
       "./reconcile": {
         "types": "./dist/reconcile.d.ts",
@@ -228,8 +228,8 @@
         "import": "./dist/tooltip-portal.js"
       },
       "./types": {
-        "types": "./dist/portable-types.d.ts",
-        "import": "./dist/portable-types.js"
+        "types": "./dist/universal-types.d.ts",
+        "import": "./dist/universal-types.js"
       },
       "./vector": {
         "types": "./dist/vector.d.ts",
