+++ packages/charts-core/src/exports.test.ts
@@ -24,10 +24,10 @@ describe('public package exports', () => {
     ).toBe(true)
   })
 
-  it('keeps the portable barrel aligned with root authoring exports', async () => {
-    const [root, portable] = await Promise.all([
+  it('keeps the universal barrel aligned with root authoring exports', async () => {
+    const [root, universal] = await Promise.all([
       import('@tanstack/charts'),
-      import('@tanstack/charts/portable'),
+      import('@tanstack/charts/universal'),
     ])
     const browserOnlyRootValues = new Set([
       'createChartAdapter',
@@ -36,7 +36,7 @@ describe('public package exports', () => {
       'resolveChartAdapterLayout',
     ])
 
-    expect(Object.keys(portable).sort()).toEqual(
+    expect(Object.keys(universal).sort()).toEqual(
       Object.keys(root)
         .filter((name) => !browserOnlyRootValues.has(name))
         .sort(),
