+++ packages/charts-core/src/index.ts
@@ -172,7 +172,7 @@ export type {
   SceneRule,
   SceneStyle,
   StaticChartDefinition,
-} from './portable-types'
+} from './universal-types'
 export { areaX } from './area-x'
 export type { AreaXCurve, AreaXOptions } from './area-x'
 export { d3AreaXCurve } from './d3-area-x'
