+++ packages/charts-core/src/universal.ts
@@ -25,4 +25,4 @@ export { ruleX, ruleY } from './rule'
 export { text } from './text'
 export { tickX, tickY } from './tick'
 export { vector } from './vector'
-export type * from './portable-types'
+export type * from './universal-types'
